requirements.txt
asgiref==3.8.1
Django==4.2.17
django-filter==24.3
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
phonenumbers==8.13.51
sqlparse==0.5.2
typing_extensions==4.12.2
